web/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
requests
python-docx
PyPDF2
pdfplumber
beautifulsoup4
lxml
psutil
pytesseract
Pillow
pdf2image